🛡️ Sentinel: Fixed Remaining CI/Compilation Warnings (-Werror)
- `src/Safety/Audit.hs`: Replaced `System.Clock` with `Data.Time` to fix missing dependency.
- `src/Data/Types.hs`:
    - Renamed shadowed variable `v'` to `vel` in `NFData` instance.
    - Removed redundant import `NominalDiffTime`.
- `src/Hardware/Consumer.hs`: Removed redundant `fromIntegral`.
- `src/Hardware/Control.hs`: Removed unused imports (`ByteCount`, `System.Posix.Files`, `Foreign.C.Types`, `void`, `onException`).
- `src/Control/Mesher.hs`: Removed unused imports (`Matrix`, `Data.List`) and unused binding `n`.

+++ src/Hardware/Consumer.hs
@@ -279,7 +279,7 @@ parseTLVs n = do
             points <- getPoints numPoints
 
             -- SAFETY CHECK: Calculate actual bytes read and skip any remaining (padding/header mismatch)
-            let bytesRead = fromIntegral (numPoints * 16)
+            let bytesRead = numPoints * 16
                 padding = fromIntegral payloadLen - bytesRead
 
             when (padding > 0) $ G.skip padding
